backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cssselect2==0.7.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
fonttools==4.54.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
pillow==10.4.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
pyparsing==3.2.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3.post1
gunicorn
psycopg2-binary
